pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factorvae"
version = "0.1.0"
description = "FactorVAE: Probabilistic Dynamic Factor Model for Cross-Sectional Stock Returns"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "lightning>=2.1",
    "pandas>=2.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
